Full stack running: Rust API + gRPC worker + Postgres + Reactome
All three services now run together and answer a real query end to end.
Imatinib at cutoff 0.35 returns 2 similar compounds (imatinib 1.000,
nilotinib 0.517) -> 3 human targets (ABL1, KIT, PDGFRA) -> 100 Reactome
pathways, in 3.6 ms. The non-human target is correctly excluded by the
organism filter.

Toolchain: installed Rust 1.97 and protoc 35.1. The Rust API compiled
clean on the first attempt -- axum 0.8 routing and the sqlx UNNEST array
binding both worked as written.

Reactome ingested: 2,883 human pathways (29 top-level), 2,899 hierarchy
edges, 142,108 protein-pathway links.

Bugs found by actually running things:

- biological_domain was NULL for all 2,883 pathways. Migration 003's
  denormalising UPDATE runs once, against a table that is still empty at
  migration time, so it never took effect. Worse, refresh_derived rebuilt
  target_pathway (which reads that column) BEFORE pathway_domain computed
  it. Refresh is now explicitly ordered and re-runs the UPDATE.
- /api/status reported a healthy database as down: `SELECT 1` is INT4 in
  Postgres and was being decoded as i64.
- /api/resolve could not match any structure to ChEMBL, because the test
  fixture used placeholder InChIKeys. Replaced with the real RDKit-computed
  keys -- resolve now returns CHEMBL941 / IMATINIB.
- The loader used ON CONFLICT DO NOTHING, so re-ingesting a newer ChEMBL
  release would silently ignore corrected structures and updated clinical
  phases forever. Now upserts, falling back to DO NOTHING only for pure
  join tables that have no non-key columns.

Added:
- gen_proto.py, which regenerates the gRPC stubs and patches grpcio's flat
  `import chemworker_pb2` to be package-relative. Without it the server
  cannot import at all.
- Migration 005 carrying ChEMBL's pref_name through to chem.compound, so
  the API can say "imatinib" rather than only "CHEMBL941".

32 tests pass, ruff clean.

+++ services/api/src/routes/health.rs
@@ -31,7 +31,9 @@ pub enum ComponentStatus {
 /// "the graph is empty" and "the worker is down" look identical in the UI
 /// otherwise.
 pub async fn status(State(state): State<AppState>) -> ApiResult<Json<StatusResponse>> {
-    let database = match sqlx::query_scalar::<_, i64>("SELECT 1")
+    // `SELECT 1` is INT4 in Postgres, not INT8 -- decoding it as i64 fails at
+    // runtime and made a perfectly healthy database report itself as down.
+    let database = match sqlx::query_scalar::<_, i32>("SELECT 1")
         .fetch_one(&state.db)
         .await
     {
